Move search utils into `searchers` module
